finished implementing farthest point, changed depth_estimator node to obstacle_detection node, and cleaned up/commented
