--- packages/autonomy/ezrassor_autonomous_control/nodes/depth_estimator
@@ -1,4 +1,0 @@
-#!/usr/bin/env python
-from ezrassor_autonomous_control import depth_estimator
-
-depth_estimator.depth_estimator()
