+++ packages/autonomy/ezrassor_autonomous_control/nodes/obstacle_detection
@@ -1,0 +1,4 @@
+#!/usr/bin/env python
+from ezrassor_autonomous_control import obstacle_detection
+
+obstacle_detection.obstacle_detection()
