+++ packages/autonomy/ezrassor_autonomous_control/source/ezrassor_autonomous_control/__init__.py
@@ -3,5 +3,5 @@ import ai_objects
 import nav_functions
 import utility_functions
 import transform_broadcaster
-import depth_estimator
+import obstacle_detection
 import auto_functions
